+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- change endnote identifierInPrimarySource hyphenation
+
 ### Security
 
 ## [0.37.1] - 2025-07-01